# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pingit API",
        "description": "Pingit is a full-stack uptime monitoring system. Monitor your URLs, get notified when they go down, and track performance over time.",
        "contact": {
            "email": "support@pingit.live"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://pingit.live",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/monitors/{monitorId}/channels": {
            "get": {
                "tags": [
                    "Notification Channels"
                ],
                "summary": "List all notification channels for a monitor",
                "operationId": "4785efd78c8d2a7ba613479344c59cfc",
                "parameters": [
                    {
                        "name": "monitorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of notification channels"
                    },
                    "404": {
                        "description": "Monitor not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notification Channels"
                ],
                "summary": "Add a notification channel to a monitor",
                "operationId": "cbc3a068df08b6e3ca4550d6d143c43b",
                "parameters": [
                    {
                        "name": "monitorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "value"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "email",
                                            "webhook"
                                        ],
                                        "example": "email"
                                    },
                                    "value": {
                                        "type": "string",
                                        "example": "alerts@example.com"
                                    },
                                    "notify_on_down": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "notify_on_recovery": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "notify_on_degraded": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notification channel created"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Monitor not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/monitors/{monitorId}/channels/{channelId}": {
            "delete": {
                "tags": [
                    "Notification Channels"
                ],
                "summary": "Delete a notification channel",
                "operationId": "1053ed8d7fb5e04cf5d48183a6196546",
                "parameters": [
                    {
                        "name": "monitorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "channelId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification channel deleted"
                    },
                    "404": {
                        "description": "Monitor or channel not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "List all tags for the authenticated user",
                "operationId": "b74c2b2170c7593429cf9963ef8b474b",
                "responses": {
                    "200": {
                        "description": "List of tags"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Create a new tag",
                "operationId": "9f1625bd2d2f18ea0d49f1fb6c105afc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "production"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tag created"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tags/{id}": {
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "Delete a tag",
                "operationId": "b4af5e1e3330a1b80a48de3e31338442",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag deleted"
                    },
                    "404": {
                        "description": "Tag not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/monitors/{monitorId}/tags": {
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Attach a tag to a monitor",
                "operationId": "0d3c4200fc3bf72dca4bd84344f706d2",
                "parameters": [
                    {
                        "name": "monitorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tag_id"
                                ],
                                "properties": {
                                    "tag_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tag attached successfully"
                    },
                    "404": {
                        "description": "Monitor or tag not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/monitors/{monitorId}/tags/{tagId}": {
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "Detach a tag from a monitor",
                "operationId": "c013ecc89f84096d8e08f9c4e52bc906",
                "parameters": [
                    {
                        "name": "monitorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tagId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag detached successfully"
                    },
                    "404": {
                        "description": "Monitor not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}